Reworked tail animation, added new molang variables

+++ src/main/java/by/dragonsurvivalteam/dragonsurvival/util/Functions.java
@@ -31,9 +31,9 @@ public class Functions{
 		return ticks / 20;
 	}
 
-	public static float angleDifference(float angle1, float angle2){
-		float phi = Math.abs(angle1 - angle2) % 360;
-		float dif = phi > 180 ? 360 - phi : phi;
+	public static double angleDifference(double angle1, double angle2){
+		double phi = Math.abs(angle1 - angle2) % 360;
+		double dif = phi > 180 ? 360 - phi : phi;
 		int sign = angle1 - angle2 >= 0 && angle1 - angle2 <= 180 || angle1 - angle2 <= -180 && angle1 - angle2 >= -360 ? 1 : -1;
 		dif *= sign;
 		return dif;
